progress

+++ Arduino Pro Mini Code/Pro_mini_code/Pro_mini_code.ino
@@ -46,7 +46,7 @@
 
 typedef union
 {
-  float number;
+  int32_t number;
   uint8_t bytes[4];
 } FLOATUNION_t;
